+++ paddle/fluid/operators/CMakeLists.txt
@@ -104,7 +104,7 @@ endif()
 set(OP_HEADER_DEPS ${OP_HEADER_DEPS} phi phi_api_utils gather_scatter_kernel backward_infermeta)
 
 register_operators(EXCLUDES py_layer_op py_func_op warpctc_op dgc_op load_combine_op lstm_op run_program_op eye_op quantize_linear_op
-        recurrent_op save_combine_op sparse_attention_op sync_batch_norm_op spectral_op ${OP_MKL_DEPS} DEPS ${OP_HEADER_DEPS})
+        recurrent_op save_combine_op sparse_attention_op sync_batch_norm_op ${OP_MKL_DEPS} DEPS ${OP_HEADER_DEPS})
 
 op_library(run_program_op SRCS run_program_op.cc run_program_op.cu.cc DEPS executor_cache ${OP_HEADER_DEPS})
 target_link_libraries(run_program_op cuda_graph_with_memory_pool)
@@ -127,22 +127,6 @@ if (WITH_GPU OR WITH_ROCM)
     endif()
 else()
     op_library(warpctc_op DEPS dynload_warpctc sequence_padding sequence_scale)
-endif()
-
-if (WITH_GPU OR WITH_ROCM)
-    if (MKL_FOUND AND WITH_ONEMKL)
-        op_library(spectral_op SRCS spectral_op.cc spectral_op.cu DEPS dynload_cuda dynload_mklrt ${OP_HEADER_DEPS})
-        target_include_directories(spectral_op PRIVATE ${MKL_INCLUDE})
-    else()
-        op_library(spectral_op SRCS spectral_op.cc spectral_op.cu DEPS dynload_cuda ${OP_HEADER_DEPS})
-    endif()
-else()
-    if (MKL_FOUND AND WITH_ONEMKL)
-        op_library(spectral_op SRCS spectral_op.cc DEPS dynload_mklrt ${OP_HEADER_DEPS})
-        target_include_directories(spectral_op PRIVATE ${MKL_INCLUDE})
-    else()
-        op_library(spectral_op SRCS spectral_op.cc DEPS ${OP_HEADER_DEPS})
-    endif()
 endif()
 
 if (WITH_ASCEND_CL)
