+++ paddle/phi/api/yaml/api.yaml
@@ -98,6 +98,33 @@
     func : erf
   backward : erf_grad
 
+- api : fft_c2c
+  args : (Tensor x, int64_t[] axes, str normalization, bool forward)
+  output : Tensor
+  infer_meta :
+    func : FFTC2CInferMeta
+  kernel :
+    func : fft_c2c
+  backward : fft_c2c_grad
+
+- api : fft_c2r
+  args : (Tensor x, int64_t[] axes, str normalization, bool forward, int64_t last_dim_size=0L)
+  output : Tensor
+  infer_meta :
+    func : FFTC2RInferMeta
+  kernel :
+    func : fft_c2r
+  backward : fft_c2r_grad
+
+- api : fft_r2c
+  args : (Tensor x, int64_t[] axes, str normalization, bool forward, bool onesided)
+  output : Tensor
+  infer_meta :
+    func : FFTR2CInferMeta
+  kernel :
+    func : fft_r2c
+  backward : fft_r2c_grad
+
 - api : lgamma
   args : (Tensor x)
   output : Tensor(out)
@@ -105,7 +132,7 @@
     func : UnchangedInferMeta
   kernel :
     func : lgamma
-  backward : lgamma_grad 
+  backward : lgamma_grad
 
 - api : mv
   args : (Tensor x, Tensor vec)
