+++ paddle/phi/api/yaml/api_compat.yaml
@@ -31,6 +31,15 @@
              float Scale_in_eltwise = 1.0f, 'float[] Scale_weights = {1.0f}', bool force_fp32_output = false,
              int workspace_size_MB = 512, bool exhaustive_search = false]
 
+- api : conv2d
+  extra :
+    attrs : [bool use_cudnn = false, bool fuse_relu_before_depthwise_conv = false, bool use_mkldnn = false,
+             bool use_quantizer = false, str mkldnn_data_type = "float32", bool fuse_relu = false,
+             str fuse_activation = "", bool fuse_alpha = false, bool fuse_beta = false, bool use_addto = false,
+             bool fuse_residual_connection = false, float Scale_in = 1.0f, float Scale_out = 1.0f,
+             float Scale_in_eltwise = 1.0f, 'float[] Scale_weights = {1.0f}', bool force_fp32_output = false,
+             int workspace_size_MB = 512, bool exhaustive_search = false]
+
 - api : cross
   inputs :
     {x : X, y : Y}
@@ -112,3 +121,15 @@
     x : X
   outputs :
     out : Out
+
+- api: fft_c2c
+  inputs: {x: X}
+  outputs: {out: Out}
+
+- api: fft_c2r
+  inputs: {x: X}
+  outputs: {out: Out}
+
+- api: fft_r2c
+  inputs: {x: X}
+  outputs: {out: Out}
