+++ paddle/phi/api/yaml/backward.yaml
@@ -105,6 +105,38 @@
     func : erf_grad
     data_type : out_grad
 
+- backward_api : fft_c2c_grad
+  forward: fft_c2c(Tensor x, int64_t[] axes, str normalization, bool forward) -> Tensor(out)
+  args : (Tensor out_grad, int64_t[] axes, str normalization, bool forward)
+  output: Tensor(x_grad)
+  infer_meta :
+    func : UnchangedInferMeta
+    param : [out_grad]
+  kernel :
+    func : fft_c2c_grad
+
+- backward_api : fft_c2r_grad
+  forward: fft_c2r(Tensor x, int64_t[] axes, str normalization, bool forward, int64_t last_dim_size) -> Tensor(out)
+  args : (Tensor out_grad, int64_t[] axes, str normalization, bool forward, int64_t last_dim_size)
+  output: Tensor(x_grad)
+  infer_meta :
+    func : FFTC2RGradInferMeta
+  kernel :
+    func : fft_c2r_grad
+    data_type: out_grad
+
+- backward_api : fft_r2c_grad
+  forward: fft_r2c(Tensor x, int64_t[] axes, str normalization, bool forward, bool onesided) -> Tensor(out)
+  args : (Tensor x, Tensor out_grad, int64_t[] axes, str normalization, bool forward, bool onesided)
+  output: Tensor(x_grad)
+  infer_meta :
+    func : UnchangedInferMeta
+    param : [x]
+  kernel :
+    func : fft_r2c_grad
+    data_type: out_grad
+  no_need_buffer: x
+
 - backward_api : lgamma_grad
   forward : lgamma(Tensor x) -> Tensor(out)
   args : (Tensor x, Tensor out_grad)
