+++ paddle/phi/kernels/cpu/scale_kernel.cc
@@ -62,4 +62,6 @@ PD_REGISTER_KERNEL(scale,
                    int8_t,
                    int16_t,
                    int,
-                   int64_t) {}
+                   int64_t,
+                   phi::dtype::complex<float>,
+                   phi::dtype::complex<double>) {}
