+++ paddle/phi/kernels/funcs/CMakeLists.txt
@@ -16,3 +16,20 @@ math_library(pooling DEPS dense_tensor)
 math_library(segment_pooling)
 math_library(sequence2batch)
 math_library(matrix_solve DEPS dense_tensor eigen3 blas math_function)
+
+if(WITH_GPU OR WITH_ROCM)
+  if(MKL_FOUND AND WITH_ONEMKL)
+    math_library(fft spectral_op.cu DEPS dynload_cuda dynload_mklrt
+                 dense_tensor)
+    target_include_directories(fft PRIVATE ${MKL_INCLUDE})
+  else()
+    math_library(fft spectral_op.cu DEPS dynload_cuda dense_tensor pocketfft)
+  endif()
+else()
+  if(MKL_FOUND AND WITH_ONEMKL)
+    mathp_library(fft DEPS dynload_mklrt dense_tensor)
+    target_include_directories(fft PRIVATE ${MKL_INCLUDE})
+  else()
+    math_library(fft DEPS dense_tensor pocketfft)
+  endif()
+endif()
