+++ paddle/phi/kernels/gpu/scale_kernel.cu
@@ -74,4 +74,6 @@ PD_REGISTER_KERNEL(scale,
                    int8_t,
                    int16_t,
                    int,
-                   int64_t) {}
+                   int64_t,
+                   phi::dtype::complex<float>,
+                   phi::dtype::complex<double>) {}
